requirements/base.txt
et-xmlfile==1.1.0
mysql-connector-python==8.0.26
numpy==1.21.1
openpyxl==3.0.7
pandas==1.3.1
protobuf==3.17.3
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
